moved `num_tokens` to `utils.py`
moved `num_tokens` to `icefall/utils.py` to reduce code redundancy
