+++ icefall/utils.py
@@ -2060,3 +2060,22 @@ def symlink_or_copy(exp_dir: Path, src: str, dst: str):
     except OSError:
         copyfile(src=exp_dir / src, dst=exp_dir / dst)
     os.close(dir_fd)
+
+def num_tokens(
+    token_table: k2.SymbolTable, disambig_pattern: str = re.compile(r"^#\d+$")
+) -> int:
+    """Return the number of tokens excluding those from
+    disambiguation symbols.
+
+    Caution:
+      0 is not a token ID so it is excluded from the return value.
+    """
+    symbols = token_table.symbols
+    ans = []
+    for s in symbols:
+        if not disambig_pattern.match(s):
+            ans.append(token_table[s])
+    num_tokens = len(ans)
+    if 0 in ans:
+        num_tokens -= 1
+    return num_tokens
